Move LSP's method pretty printer to shared lib
