+++ main/lsp/requests/hover.cc
@@ -3,6 +3,7 @@
 #include "absl/strings/str_join.h"
 #include "common/sort/sort.h"
 #include "core/lsp/QueryResponse.h"
+#include "core/lsp/helpers.h"
 #include "main/lsp/LSPLoop.h"
 #include "main/lsp/LSPQuery.h"
 #include "main/lsp/json_types.h"
@@ -23,8 +24,8 @@ string methodInfoString(const core::GlobalState &gs, const core::TypePtr &retTyp
             if (!contents.empty()) {
                 contents += "\n";
             }
-            contents = absl::StrCat(
-                contents, prettyTypeForMethod(gs, component.method, component.receiver, retType, constraint.get()));
+            contents = absl::StrCat(contents, core::lsp::prettyTypeForMethod(gs, component.method, component.receiver,
+                                                                             retType, constraint.get()));
         }
         start = start->secondary.get();
     }
@@ -113,7 +114,7 @@ unique_ptr<ResponseMessage> HoverTask::runRequest(LSPTypecheckerDelegate &typech
             typeString = methodInfoString(gs, retType, *sendResp->dispatchResult, constraint);
         }
     } else if (auto defResp = resp->isMethodDef()) {
-        typeString = prettyTypeForMethod(gs, defResp->symbol, nullptr, defResp->retType.type, nullptr);
+        typeString = core::lsp::prettyTypeForMethod(gs, defResp->symbol, nullptr, defResp->retType.type, nullptr);
     } else if (auto constResp = resp->isConstant()) {
         typeString = prettyTypeForConstant(gs, constResp->symbol);
     } else {
