+++ main/lsp/requests/signature_help.cc
@@ -1,5 +1,6 @@
 #include "main/lsp/requests/signature_help.h"
 #include "core/lsp/QueryResponse.h"
+#include "core/lsp/helpers.h"
 #include "main/lsp/LSPLoop.h"
 #include "main/lsp/LSPQuery.h"
 #include "main/lsp/json_types.h"
@@ -37,8 +38,8 @@ void addSignatureHelpItem(const core::GlobalState &gs, core::MethodRef method,
         if (i != args.size() - 1) {
             methodDocumentation += ", ";
         }
-        parameter->documentation = getResultType(gs, arg.type, method, resp.dispatchResult->main.receiver,
-                                                 resp.dispatchResult->main.constr.get())
+        parameter->documentation = core::lsp::getResultType(gs, arg.type, method, resp.dispatchResult->main.receiver,
+                                                            resp.dispatchResult->main.constr.get())
                                        .show(gs);
         parameters.push_back(move(parameter));
         i += 1;
